change column type decimal to float

+++ SanalBorsaAPI.Core/Entities/CryptoCurrency.cs
@@ -11,8 +11,8 @@ namespace SanalBorsaAPI.Core.Entities
         public int Id { get; set; }
         public string Name { get; set; }
         public string ShortName { get; set; }
-        public decimal BuyPriceDolar { get; set; }
-        public decimal BuyPriceTL { get; set; }
+        public float BuyPriceDolar { get; set; }
+        public float BuyPriceTL { get; set; }
         public string MarketingValue { get; set; }
         public string MarketingSize { get; set; }
         public decimal Change { get; set; }
