+++ SanalBorsaAPI.Core/Entities/Stocks.cs
@@ -11,9 +11,9 @@ namespace SanalBorsaAPI.Core.Entities
         public int Id { get; set; }
         public string Name { get; set; }
         public string ShortName { get; set; }
-        public decimal LastPrice { get; set; }
-        public decimal HighestPrice { get; set; }
-        public decimal LowestPrice { get; set; }
+        public float LastPrice { get; set; }
+        public float HighestPrice { get; set; }
+        public float LowestPrice { get; set; }
         public string MarketingSize { get; set; }
         public decimal Change { get; set; }
         public DateTime UpdateTime { get; set; }
